requirements.txt
requests>=2.25.0
pillow>=11.1.0
pystray>=0.19.5
pytz>=2025.1
pyinstaller==6.12.0